pyproject.toml
[tool.poetry]
name = "b2upload"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^7.0.0"
pycairo = "^1.19.0"
PyGObject = "^3.34.0"
toml = "^0.10.0"
b2sdk = "^1.0.2"
pyperclip = "^1.7.0"
xxhash = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
